modify the code, and  try to solve WD triggered problem

+++ include/AGIS_Commons.h
@@ -6,7 +6,7 @@
 extern volatile unsigned int numDrops;
 extern volatile unsigned int dripRate;
 extern volatile unsigned int infusedVolume_x100;
-extern volatile unsigned long infusedTime;
+extern volatile unsigned int infusedTime;
 extern unsigned int dropFactor;
 
 extern unsigned int targetVTBI;
